Move id validation method to utils file
