+++ src/components/App/App.js
@@ -5,38 +5,13 @@ import PhotosContainer from '../PhotosContainer/PhotosContainer';
 import Loader from '../Loader/Loader';
 import Error from '../Error/Error';
 import React, { useState, useEffect } from 'react'
-import { getPhotosByAlbum } from '../../utils';
+import { getPhotosByAlbum, validateId } from '../../utils';
 
 const App = () => {
   const [isLoading, setIsLoading] = useState(false);
   const [isError, setIsError] = useState(false);
   const [currentAlbumNumber, setCurrentAlbumNumber] = useState('');
   const [photos, setPhotos] = useState([]);
-
-  const validateId = (id) => {
-    let splitId = id.split('')
-    if (splitId.length === 2 && splitId[0] === '0' && splitId[1] !== '0') {
-      splitId.shift()
-      let result = splitId.join()
-      return result
-    } else if (splitId.length === 3 && splitId[0] === '0' && splitId[1] === '0' && splitId[2] === '0') {
-      return '0'
-    } else if (splitId.length === 3 && splitId[0] === '0' && splitId[1] === '0' && splitId[2] !== '0') {
-      splitId.splice(0,2) 
-      let result = splitId.join()
-      return result
-    } else if (splitId.length === 3 && splitId[0] === '0' && splitId[1] !== '0' && splitId[2] !== '0') {
-      splitId.shift()
-      let result = splitId.join('')
-      return result
-    } else if (splitId.length === 3 && splitId[0] === '0' && splitId[1] !== '0' && splitId[2] === '0') {
-      splitId.shift()
-      let result = splitId.join('')
-      return result
-    } else {
-      return id
-    }
-  }
 
   const changeCurrentAlbumNumber = (id) => {
     let validID = validateId(id)
